==> src/05_Clusters-humancellmap.Rmd <==
---
title: "05: Clusters Human Cell Map"
output: 
  html_document:
    toc: true
    toc_float: true
    theme: flatly
---

# Setup 

Remove all objects from the workspace
```{r clear-workspace, message=FALSE, warning=FALSE}
rm(list = ls())
```

Libraries required
```{r libraries, message=FALSE, warning=FALSE}
library(here)
library(DEP)
library(dplyr)
library(ggplot2)
library(tidyr)
```

Prepare a folder for the outputs
```{r outputs-folder, message=FALSE, warning=FALSE}
if (!dir.exists(here("outputs", "05_Clusters_HCM"))) {
  dir.create(here("outputs", "05_Clusters_HCM"))
}
```

# Data input

```{r data-input, message=FALSE, warning=FALSE}
load(here("outputs",  "01_DEP-processing", "01_DEP-heatmap-clusters.RData"))
load(here("outputs", "01_DEP-processing", "01_DEP-results-table.RData"))
hcm <- read.delim(here("data", "preys-latest.txt"))
```

# Preprocess the data

```{r data-preprocess, message=FALSE, warning=FALSE}
# Significantly changed interactors of baits
data_results <- data_results %>%
  mutate(up_DVL3 = ifelse(DVL3_vs_CTRL_p.adj < 0.05 & DVL3_vs_CTRL_ratio > 1, 1, 0)) %>%
  mutate(up_ROR1 = ifelse(ROR1_vs_CTRL_p.adj < 0.05 & ROR1_vs_CTRL_ratio > 1, 1, 0)) %>%
  mutate(up_ROR2 = ifelse(ROR2_vs_CTRL_p.adj < 0.05 & ROR2_vs_CTRL_ratio > 1, 1, 0)) %>%
  mutate(up_PRICKLE1 = ifelse(PRICKLE1_vs_CTRL_p.adj < 0.05 & PRICKLE1_vs_CTRL_ratio > 1, 1, 0)) %>%
  mutate(up_VANGL2 = ifelse(VANGL2_vs_CTRL_p.adj < 0.05 & VANGL2_vs_CTRL_ratio > 1, 1, 0))

clusters <- data.heatmap %>%
  left_join(., hcm %>%
              select(symbol, MMF.localization),
            by = c("protein" = "symbol")) %>%
  left_join(., data_results %>%
              select(name, starts_with("up_")),
            by = c("protein" = "name"))

# Create long table
clusters <- clusters %>%
  select(protein, starts_with("up_"), MMF.localization, k) %>%
  rename_with(~ sub("up_", "", .x), everything()) %>%
  pivot_longer(cols = DVL3:VANGL2,
               names_to = "bait",
               values_to = "significant")
```

# Plot Human Cell Map for the clusters

```{r plot-HCM, message=FALSE, warning=FALSE}
# Significantly changed interactors of baits
#svg(here("outputs", "05_Clusters_HCM", "05_clusters-HCM.svg"))
clusters %>%
  na.omit() %>%
  group_by(k, MMF.localization) %>%
  summarise(n_preys = sum(significant)) %>%
  filter(n_preys != 0) %>%
  mutate(top_loc = case_when(
                             n_preys == max(n_preys) ~ "max",
                             TRUE ~ "")
  ) %>%
  #mutate(cluster = factor(k, levels = c(2,3,4,1,5))) %>%
  mutate(cluster = factor(k, levels = c(1, 2, 3, 4, 5), labels = c("#C1", "#C2", "#C3", "#C4", "#C5"))) %>%
  ggplot(aes(x = MMF.localization, y = n_preys, fill = top_loc)) +
  geom_bar(stat = "identity") +
  theme_minimal() +
  theme(axis.text.x = element_text(angle = 45, hjust = 1)) +
  theme(legend.position = "none") +
  labs(x = "Prey localization in Human Cell Map",
       y = "Number of preys") +
  facet_grid(cluster ~ .) +
  scale_fill_manual(values = c("#b8B8B8", "#E64B35FF"))
#dev.off()
```
